webapp/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
kubernetes>=29.0.0
redis>=5.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
pyyaml>=6.0
python-multipart>=0.0.9
httpx>=0.27.0
